+++ packages/web-components/src/components/feedback/__stories-dont-show__/feedback.stories.js
@@ -7,8 +7,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 import '../components/feedback/feedback';
-import Flag24 from '@carbon/web-components/es/icons/flag/24.js';
+import Flag24 from '@carbon/icons/es/flag/24.js';
 import { html } from 'lit';
+import { iconLoader } from '@carbon/web-components/es/globals/internal/icon-loader.js';
 
 export default {
   title: 'Components/Feedback/Feedback',
@@ -165,7 +166,7 @@ export const Default = {
         generation-id=${args['generation-id']}
         .feedbackList=${args['feedback-list']}
         icon=${args.icon}>
-        ${Flag24({ slot: 'icon' })} ${args.output}
+        ${iconLoader(Flag24, { slot: 'icon' })} ${args.output}
       </clabs-feedback>
     `,
 };
@@ -207,7 +208,7 @@ export const Playground = {
         generation-id=${args['generation-id']}
         .feedbackList=${args['feedback-list']}
         icon=${args.icon}>
-        ${Flag24({ slot: 'icon' })} ${args.output}
+        ${iconLoader(Flag24, { slot: 'icon' })} ${args.output}
       </clabs-feedback>
     `,
 };