+++ packages/web-components/src/components/global-header/components/global-header/src/globals/__tests__/useScript.test.ts
@@ -31,7 +31,7 @@ const propsWithAssisitMeConfig: HeaderProps = {
   ],
   profileFooterLinks: [
     {
-      text: 'Logout',
+      text: 'Log out',
       href: '/logout',
       carbonIcon: 'Logout',
       arialLabel: 'Logout',
@@ -186,7 +186,7 @@ const propsNoAssisitMe: HeaderProps = {
   ],
   profileFooterLinks: [
     {
-      text: 'Logout',
+      text: 'Log out',
       href: '/logout',
       carbonIcon: 'Logout',
       arialLabel: 'Logout',
@@ -331,7 +331,7 @@ describe('useScript function', () => {
     const status = useScript.default(propsNoAssisitMe);
     expect(status).to.equal('idle');
   });
-
+  // this test is failing with .only (I expect it's relient on earlier test).
   it('should return status of "loading" or "ready" when script is in DOM an function is called', async () => {
     const el = await fixture(html`
       <script