+++ packages/react/CHANGELOG.md
@@ -3,6 +3,114 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.148.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.147.0...@carbon-labs/react@0.148.0) (2025-11-14)
+
+
+### Bug Fixes
+
+* **uishell:** update header overflow action button styles ([#914](https://github.com/carbon-design-system/carbon-labs/issues/914)) ([83eb662](https://github.com/carbon-design-system/carbon-labs/commit/83eb662bc25a4d454999119761ccce85c49b1d61))
+
+
+
+
+
+# [0.147.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.146.0...@carbon-labs/react@0.147.0) (2025-11-10)
+
+
+### Features
+
+* **uishell:** update HeaderPopover styles ([#890](https://github.com/carbon-design-system/carbon-labs/issues/890)) ([2543742](https://github.com/carbon-design-system/carbon-labs/commit/254374236689fb8e338b493bd5c4ff1927c7e23d))
+
+
+
+
+
+# [0.146.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.145.0...@carbon-labs/react@0.146.0) (2025-11-07)
+
+
+### Bug Fixes
+
+* **Text-Highlighter:** Overview page error fix, story and example tweak ([#905](https://github.com/carbon-design-system/carbon-labs/issues/905)) ([ddd0716](https://github.com/carbon-design-system/carbon-labs/commit/ddd0716d99650eca3ef5d27d4828f8ab0d5ca59b))
+
+
+
+
+
+# [0.145.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.144.0...@carbon-labs/react@0.145.0) (2025-11-03)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.144.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.143.0...@carbon-labs/react@0.144.0) (2025-10-30)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.143.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.142.0...@carbon-labs/react@0.143.0) (2025-10-30)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.142.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.141.0...@carbon-labs/react@0.142.0) (2025-10-24)
+
+
+### Features
+
+* **header-overflow-panel:** add to ui-shell ([#854](https://github.com/carbon-design-system/carbon-labs/issues/854)) ([f2f1f61](https://github.com/carbon-design-system/carbon-labs/commit/f2f1f618ed94e679cc10aa1cda0107d246fc11ed))
+
+
+
+
+
+# [0.141.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.140.0...@carbon-labs/react@0.141.0) (2025-10-22)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.140.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.139.0...@carbon-labs/react@0.140.0) (2025-10-22)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.139.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.138.0...@carbon-labs/react@0.139.0) (2025-10-20)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.138.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.137.0...@carbon-labs/react@0.138.0) (2025-10-16)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.137.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.136.0...@carbon-labs/react@0.137.0) (2025-10-15)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
 # [0.136.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.135.0...@carbon-labs/react@0.136.0) (2025-10-10)