+++ packages/react/src/components/StylePicker/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.6.0...@carbon-labs/react-style-picker@0.7.0) (2025-10-31)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.6.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.5.0...@carbon-labs/react-style-picker@0.6.0) (2025-10-22)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
+# [0.5.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.4.0...@carbon-labs/react-style-picker@0.5.0) (2025-10-15)
+
+**Note:** Version bump only for package @carbon-labs/react-style-picker
+
+
+
+
+
 # [0.4.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-style-picker@0.3.0...@carbon-labs/react-style-picker@0.4.0) (2025-09-22)
 
 **Note:** Version bump only for package @carbon-labs/react-style-picker