+++ packages/react/src/components/UIShell/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.87.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.86.0...@carbon-labs/react-ui-shell@0.87.0) (2025-11-14)
+
+
+### Bug Fixes
+
+* **uishell:** update header overflow action button styles ([#914](https://github.com/carbon-design-system/carbon-labs/issues/914)) ([83eb662](https://github.com/carbon-design-system/carbon-labs/commit/83eb662bc25a4d454999119761ccce85c49b1d61))
+
+
+
+
+
+# [0.86.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.85.0...@carbon-labs/react-ui-shell@0.86.0) (2025-11-10)
+
+
+### Features
+
+* **uishell:** update HeaderPopover styles ([#890](https://github.com/carbon-design-system/carbon-labs/issues/890)) ([2543742](https://github.com/carbon-design-system/carbon-labs/commit/254374236689fb8e338b493bd5c4ff1927c7e23d))
+
+
+
+
+
+# [0.85.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.84.0...@carbon-labs/react-ui-shell@0.85.0) (2025-11-03)
+
+**Note:** Version bump only for package @carbon-labs/react-ui-shell
+
+
+
+
+
+# [0.84.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.83.0...@carbon-labs/react-ui-shell@0.84.0) (2025-10-30)
+
+**Note:** Version bump only for package @carbon-labs/react-ui-shell
+
+
+
+
+
+# [0.83.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.82.0...@carbon-labs/react-ui-shell@0.83.0) (2025-10-24)
+
+
+### Features
+
+* **header-overflow-panel:** add to ui-shell ([#854](https://github.com/carbon-design-system/carbon-labs/issues/854)) ([f2f1f61](https://github.com/carbon-design-system/carbon-labs/commit/f2f1f618ed94e679cc10aa1cda0107d246fc11ed))
+
+
+
+
+
+# [0.82.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.81.0...@carbon-labs/react-ui-shell@0.82.0) (2025-10-15)
+
+**Note:** Version bump only for package @carbon-labs/react-ui-shell
+
+
+
+
+
 # [0.81.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-ui-shell@0.80.0...@carbon-labs/react-ui-shell@0.81.0) (2025-10-10)