+++ packages/web-components/src/components/feedback/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.24.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-ai-feedback@0.23.0...@carbon-labs/wc-ai-feedback@0.24.0) (2025-10-22)
+
+**Note:** Version bump only for package @carbon-labs/wc-ai-feedback
+
+
+
+
+
+# [0.23.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-ai-feedback@0.22.0...@carbon-labs/wc-ai-feedback@0.23.0) (2025-10-15)
+
+**Note:** Version bump only for package @carbon-labs/wc-ai-feedback
+
+
+
+
+
 # [0.22.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-ai-feedback@0.21.0...@carbon-labs/wc-ai-feedback@0.22.0) (2025-09-11)
 
 **Note:** Version bump only for package @carbon-labs/wc-ai-feedback