+++ packages/web-components/src/components/tag/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.16.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-ai-tag@0.15.0...@carbon-labs/wc-ai-tag@0.16.0) (2025-10-22)
+
+**Note:** Version bump only for package @carbon-labs/wc-ai-tag
+
+
+
+
+
+# [0.15.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-ai-tag@0.14.0...@carbon-labs/wc-ai-tag@0.15.0) (2025-10-15)
+
+**Note:** Version bump only for package @carbon-labs/wc-ai-tag
+
+
+
+
+
 # [0.14.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-ai-tag@0.13.0...@carbon-labs/wc-ai-tag@0.14.0) (2025-09-11)
 
 **Note:** Version bump only for package @carbon-labs/wc-ai-tag